Change @BukkitPluginRef contract for Cauldron
This is a required FML change for fulfilling the contract of the @BukkitPluginRef annotation in Cauldron.

The BukkitProxy class being a blank interface did not make any sense - it instructed users to downcast, but as no class implements BukkitProxy, there is nothing to downcast to.
It is replaced by a wrapper class, filled at the same time as the @Instance annotations, that can be filled by Cauldron at a later time (as Bukkit plugins cannot be initialized until all Forge mods have been).